Minor cleanup.
- Rename interactor class.
- Remove unused params.
- Change ids to Long.

+++ shared/domain/seasons/api/src/commonMain/kotlin/com/thomaskioko/tvmaniac/seasons/api/model/SeasonUiModel.kt
@@ -5,6 +5,6 @@ data class SeasonUiModel(
     val tvShowId: Long,
     val name: String,
     val overview: String,
-    val seasonNumber: Int,
+    val seasonNumber: Long,
     val episodeCount: Int
 )
